feat: move download to huggingface.rs
